Switch back to always returning a tuple
This does make pattern matching easier since we can potentially return
errors, and there is precedent for it with the Time and NaiveDateTime
types.

I've also cleaned up the examples and code to consistently use the full
`socket_address` name, which is also based on community conventions. The
additional tests now bring us up to 100% test coverage.